move authorizationRequestList to bh
