+++ src/views/repository/Settings.vue
@@ -60,7 +60,7 @@ import EditProfileForm from '@/components-v1/user/EditProfileForm';
 import EditRepositoryForm from '@/components-v1/repository/EditRepositoryForm';
 import SetAuthorizationRoleForm from '@/components-v1/repository/SetAuthorizationRoleForm';
 import AuthorizationsList from '@/components-v1/repository/AuthorizationsList';
-import AuthorizationRequestsList from '@/components-v1/repository/AuthorizationRequestsList';
+import AuthorizationRequestsList from '@/components/repository/AuthorizationRequestsList';
 import LoginForm from '@/components-v1/auth/LoginForm';
